% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbind.R
\name{fbind}
\alias{fbind}
\title{Bind two factors}
\usage{
fbind(fac, fac2)
}
\arguments{
\item{fac}{factor}

\item{fac2}{factor}
}
\value{
factor
}
\description{
Create a new factor from two exisiting factors, where the new factor's levels
are the union of the levels of the input factors.  See example of fac and fac2
that are automatically loaded with the package.
}
\examples{
fac <- factor(c("is", "the", "function"))
fac2 <- factor("working")
fbind(fac, fac2)
}
